Use cached drawing instead of direct back drawing
Traditionally, vim-cocoa use a NSImage in VIMTextView as a back
buffer, everything operations called by upper-level vim (gui_mch_
{clear_all, clear_block, insert_lines, delete_lines, draw_string,
...}) will be instantly drawn on that NSImage, then VIMTextView
will simply copy that image on -drawRect:. Thus, a simple double
buffer drawing mechanism.

However, each time we draw on that NSImage, we will have to call
-lockFocus and -unlockFocus, which is a cheap operation prior to
Mac OS X 10.6. In Mac OS X 10.6, NSImage is changed substantially,
it's no longer backed by an off-screen view, the direct impact to
vim-cocoa is, NSImage -lockFocus/-unlockFocus became much more
expensive. That's why we need to cache all the drawing operations
in batch, and only call lockFocus/unlockFocus once we need to
flush the drawing queue.

That's basically what we did in this commit: use a drawing queue
to cache all the operations, wait until gui_mac_redraw() to flush
out all these operations in batch. We hope it can be a pretty
useful optimization for 10.6, and will increase the performance in
10.5 a little bit.