Hack to make it launch on 10.9
OS X 10.9 deprecated Process Manager Carbon API [1] and stopped
passing -psn_ argument to processes when they are launched by
Finder (or the "open" command) [2], as a result we can no longer
use the trick [3] to detect whether we want to start Vim in gui
mode or not. Since all the Process Manager APIs are deprecated
I had to use another trick from [4] to launch as GUI when the
parent process name is "launchd". It's not perfect but since in
most cases you only want to launch as Terminal mode when the
parent process is a shell (bash for instance).

[1] https://twitter.com/rosyna/status/394845369269358592
[2] http://debbugs.gnu.org/cgi/bugreport.cgi?bug=15776
[3] http://stackoverflow.com/a/17682067
[4] http://lists.apple.com/archives/cocoa-dev/2006/Jun/msg01622.html